move more tests under test2 (non-grunt) folder
