+++ test2/unittests/tools/update/npmUpdater.js
